backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
neo4j==5.19.0
pydantic==2.7.4
pydantic-settings==2.2.1
kubernetes==28.1.0
PyYAML==6.0.1
python-multipart==0.0.9
